fixup: rename internal/header to internal/flagdmeta
