+++ providers/flagd/internal/flagdmeta/header.go
@@ -1,5 +1,5 @@
-// Package header defines metadata header keys shared by the flagd resolvers.
-package header
+// Package flagdmeta defines metadata header keys shared by the flagd resolvers.
+package flagdmeta
 
 // Selector is the metadata header used to pass the flag-set selector to flagd.
 const Selector = "flagd-selector"
