+++ providers/flagd/pkg/service/in_process/service_selector_header_test.go
@@ -10,7 +10,7 @@ import (
 
 	"buf.build/gen/go/open-feature/flagd/grpc/go/flagd/sync/v1/syncv1grpc"
 	v1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/flagd/sync/v1"
-	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/header"
+	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/flagdmeta"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
@@ -181,7 +181,7 @@ func (s *selectorHeaderCapturingServer) GetMetadata(ctx context.Context, req *v1
 func (s *selectorHeaderCapturingServer) captureHeader(ctx context.Context) {
 	md, _ := metadata.FromIncomingContext(ctx)
 	headerValue := ""
-if values := md.Get(header.Selector); len(values) > 0 {
+if values := md.Get(flagdmeta.Selector); len(values) > 0 {
 		headerValue = values[0]
 	}
 	select {
