+++ providers/flagd/pkg/service/rpc/interceptors.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"connectrpc.com/connect"
-	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/header"
+	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/flagdmeta"
 )
 
 // selectorInterceptor is a connect.Interceptor that adds the flagd-selector header.
@@ -19,7 +19,7 @@ func newSelectorInterceptor(selector string) *selectorInterceptor {
 func (i *selectorInterceptor) WrapUnary(next connect.UnaryFunc) connect.UnaryFunc {
 	return func(ctx context.Context, req connect.AnyRequest) (connect.AnyResponse, error) {
 		if i.selector != "" {
-			req.Header().Set(header.Selector, i.selector)
+			req.Header().Set(flagdmeta.Selector, i.selector)
 		}
 		return next(ctx, req)
 	}
@@ -29,7 +29,7 @@ func (i *selectorInterceptor) WrapStreamingClient(next connect.StreamingClientFu
 	return func(ctx context.Context, spec connect.Spec) connect.StreamingClientConn {
 		conn := next(ctx, spec)
 		if i.selector != "" {
-			conn.RequestHeader().Set(header.Selector, i.selector)
+			conn.RequestHeader().Set(flagdmeta.Selector, i.selector)
 		}
 		return conn
 	}
